[VT2-143] refactor: cn 유틸 함수를 utils/cn.ts로 구조 변경
