+++ src/components/Badge/Badge.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { cn } from '../../lib/utils'
+import { cn } from '../../utils/cn'
 import { badgeVariants, type BadgeVariantProps } from './badgeVariants'
 
 interface BadgeProps extends React.HTMLAttributes<HTMLDivElement>, BadgeVariantProps {
